refactor(tui): split TuiSession into its own module
rinkaku-tui/src/lib.rs was in ADR 0028's file-size watch band
(1212 lines, threshold 1000-1500). TuiSession (terminal lifecycle:
init/draw_splash/run, the Drop safety net) and the crate's run()
entry point are a self-contained responsibility distinct from
run_app's pure dispatch/translate glue, so they move to a new
session.rs sibling module along the same seam lib.rs's own module
doc comment already draws between the "view-model" and "terminal
adapter" layers.

Purely mechanical: no behavior change, no API change. run_app
becomes pub(crate) so session.rs can call into it; TuiSession/run
are re-exported from the crate root (pub use session::{TuiSession,
run}) so every existing caller path (main.rs's
`rinkaku_tui::TuiSession`) is unaffected.

lib.rs: 1212 -> 1038 lines (still in the watch band, but well clear
of the 1500-line warn threshold).
